报名活动建群增加：建群后是否对报名人自动发出邀请功能

+++ lbs_web/src/main/java/com/innovate/cms/modules/qs/web/push/QxPushInfoController.java
@@ -512,7 +512,7 @@ public class QxPushInfoController extends BaseController {
 			return info;
 		}
 		try {
-			long massPushId = Integer.parseInt(massPushIdStr);
+			long massPushId = Long.parseLong(massPushIdStr);
 			// 获取消息报名通知用户列表
 			List<NoticeUserForService> data = qxPushInfoService.msgNoticeUsers(massPushId);
 			backInfo.setData(data);
